+++ pkg/scalers/azure_servicebus_scaler_test.go
@@ -12,6 +12,7 @@ const (
 	queueName         = "testqueue"
 	connectionSetting = "none"
 	namespaceName     = "ns"
+	messageCount      = "1000"
 )
 
 type parseServiceBusMetadataTestData struct {
@@ -36,8 +37,12 @@ var parseServiceBusMetadataDataset = []parseServiceBusMetadataTestData{
 	{map[string]string{}, true, none, map[string]string{}, ""},
 	// properly formed queue
 	{map[string]string{"queueName": queueName, "connectionFromEnv": connectionSetting}, false, queue, map[string]string{}, ""},
+	// properly formed queue with message count
+	{map[string]string{"queueName": queueName, "connectionFromEnv": connectionSetting, "messageCount": messageCount}, false, queue, map[string]string{}, ""},
 	// properly formed topic & subscription
 	{map[string]string{"topicName": topicName, "subscriptionName": subscriptionName, "connectionFromEnv": connectionSetting}, false, subscription, map[string]string{}, ""},
+	// properly formed topic & subscription with message count
+	{map[string]string{"topicName": topicName, "subscriptionName": subscriptionName, "connectionFromEnv": connectionSetting, "messageCount": messageCount}, false, subscription, map[string]string{}, ""},
 	// queue and topic specified
 	{map[string]string{"queueName": queueName, "topicName": topicName, "connectionFromEnv": connectionSetting}, true, none, map[string]string{}, ""},
 	// queue and subscription specified
@@ -58,7 +63,7 @@ var parseServiceBusMetadataDataset = []parseServiceBusMetadataTestData{
 
 var azServiceBusMetricIdentifiers = []azServiceBusMetricIdentifier{
 	{&parseServiceBusMetadataDataset[1], "azure-servicebus-testqueue"},
-	{&parseServiceBusMetadataDataset[2], "azure-servicebus-testtopic-testsubscription"},
+	{&parseServiceBusMetadataDataset[3], "azure-servicebus-testtopic-testsubscription"},
 }
 
 var getServiceBusLengthTestScalers = []azureServiceBusScaler{